requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.0
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
